[WOOR-198] refactor: util 클래스 common 패키지로 이동
