+++ src/main/java/com/musseukpeople/woorimap/common/validator/ImageUrlValidator.java
@@ -3,6 +3,8 @@ package com.musseukpeople.woorimap.common.validator;
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
 
+import com.musseukpeople.woorimap.common.util.ImageUtil;
+
 public class ImageUrlValidator implements ConstraintValidator<ImageUrl, String> {
 
     @Override
